move files to src/pro
