+++ src/components/AgentConsentToast.tsx
@@ -143,4 +143,3 @@ export function AgentConsentToast({
     </div>
   );
 }
-
