+++ src/components/chat/DyadDatabaseSchema.tsx
@@ -11,7 +11,10 @@ interface DyadDatabaseSchemaProps {
   children: React.ReactNode;
 }
 
-export function DyadDatabaseSchema({ node, children }: DyadDatabaseSchemaProps) {
+export function DyadDatabaseSchema({
+  node,
+  children,
+}: DyadDatabaseSchemaProps) {
   const { state } = node.properties;
   const isLoading = state === "pending";
   const content = typeof children === "string" ? children : "";
@@ -34,4 +37,3 @@ export function DyadDatabaseSchema({ node, children }: DyadDatabaseSchemaProps) 
     </div>
   );
 }
-
