+++ src/components/chat/DyadToolCall.tsx
@@ -43,4 +43,3 @@ export function DyadToolCall({ node, children }: DyadToolCallProps) {
     </div>
   );
 }
-
