+++ src/hooks/useAgentTools.ts
@@ -4,7 +4,7 @@
 
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { IpcClient } from "@/ipc/ipc_client";
-import type { AgentToolName } from "@/ipc/handlers/local_agent/tool_definitions";
+import type { AgentToolName } from "../pro/main/ipc/handlers/local_agent/tool_definitions";
 import type { AgentTool } from "@/ipc/ipc_types";
 import type { AgentToolConsent } from "@/ipc/ipc_types";
