+++ src/main.ts
@@ -28,7 +28,7 @@ import {
   startPerformanceMonitoring,
   stopPerformanceMonitoring,
 } from "./utils/performance_monitor";
-import { cleanupOldAiMessagesJson } from "./ipc/handlers/local_agent/local_agent_handler";
+import { cleanupOldAiMessagesJson } from "./pro/main/ipc/handlers/local_agent/local_agent_handler";
 import fs from "fs";
 
 log.errorHandler.startCatching();
