+++ src/pro/main/ipc/handlers/local_agent/tools/search_replace.ts
@@ -3,7 +3,7 @@ import { ToolDefinition, ToolExecuteContext, escapeXmlAttr } from "./types";
 import {
   executeSearchReplaceFile,
   type FileOperationContext,
-} from "../../../processors/file_operations";
+} from "../processors/file_operations";
 
 const searchReplaceSchema = z.object({
   path: z.string().describe("The file path to edit"),
