+++ src/pro/main/ipc/handlers/local_agent/tools/set_chat_summary.ts
