+++ src/pro/main/ipc/handlers/local_agent/tools/write_file.ts
@@ -3,7 +3,7 @@ import { ToolDefinition, ToolExecuteContext, escapeXmlAttr } from "./types";
 import {
   executeWriteFile,
   type FileOperationContext,
-} from "../../../processors/file_operations";
+} from "../processors/file_operations";
 
 const writeFileSchema = z.object({
   path: z.string().describe("The file path relative to the app root"),
